restructure the way you run the program :dragon: :star:
